requirements.txt
# Существующие Jesse зависимости
jesse

# ИИ анализ
openai>=1.0.0
anthropic  # на случай если захотите Claude

# Telegram бот
python-telegram-bot>=20.0

# Утилиты
aiohttp>=3.8.0
asyncio-mqtt
redis>=4.5.0
pydantic>=2.0.0

# Дополнительная аналитика
pandas>=1.5.0
numpy>=1.24.0
talib  # технические индикаторы

# Логирование и мониторинг
structlog
sentry-sdk
